try_propose on process_qc/process_tc
try_propose should be called whenever the round is incremented. Now that
process_qc and process_tc are consolidated, we should call try_propose
in those functions.

The existing invocation in handle_vote_message is now superfluous, but
it's left in anyways so the code reads clearer. This is safe because
try_propose is idempotent.

Concretely, this change fixes suboptimal behavior where a node would not
try proposing when it bumped its round via an AdvanceRound message.